pyproject.toml
[tool.poetry]
name = "hivemind"
version = "0.1.0"
description = ""
authors = ["Mengyang (John) Cha <15233508+solarapparition@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
langchain = ">=0.0.235"
ruamel-yaml = "^0.17.32"
openai = "^0.27.8"
llama-index = "^0.8.4"
weaviate-client = "^3.22.1"
nltk = "^3.8.1"
pyautogen = {version = "^0.1.3", extras = ["blendsearch"]}

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
